Placeholder for Davide + Si

+++ mqtt-sn-codec/src/main/java/org/slj/mqtt/sn/codec/AbstractMqttsnMessageFactory.java
@@ -43,7 +43,7 @@ public abstract class AbstractMqttsnMessageFactory implements IMqttsnMessageFact
     }
 
     @Override
-    public IMqttsnMessage createIntegrityMessage(short protectionScheme, byte[] senderId, long nonce, int montonicCounter, long keyMaterial, byte[] encapsulatedMessage) throws MqttsnCodecException {
+    public IMqttsnMessage createIntegrityMessage(byte protectionScheme, byte[] senderId, long nonce, int montonicCounter, long keyMaterial, byte[] encapsulatedMessage) throws MqttsnCodecException {
         throw new MqttsnCodecException("message not supported by codec");
     }
